+++ dwengo_backend/routes/learningObject/learningObjectRoutes.ts
@@ -8,7 +8,11 @@ import {
 } from "../../controllers/learningObject/learningObjectController";
 import { protectAnyUser } from "../../middleware/authMiddleware/authAnyUserMiddleware";
 import { protectTeacher } from "../../middleware/authMiddleware/teacherAuthMiddleware";
-import { learningObjectParamsSchema } from "../../zodSchemas/paramSchemas";
+import {
+  learningObjectIdSchema,
+  learningObjectParamsSchema,
+  learningPathIdSchema,
+} from "../../zodSchemas";
 import { validateRequest } from "../../middleware/validateRequest";
 
 const router: Router = express.Router();
@@ -54,7 +58,15 @@ router.get(
  * DON'T USE THIS ROUTE: SINCE THE DWENGO API DOESN'T IMPLEMENT SEARCHING BY ID CORRECTLY,
  * THIS ROUTE ONLY CHECKS LOCAL LEARNING OBJECTS. (it's kept in in case the Dwengo API is fixed in the future)
  */
-router.get("/:learningObjectId", protectAnyUser, getLearningObjectController);
+router.get(
+  "/:learningObjectId",
+  protectAnyUser,
+  validateRequest({
+    customErrorMessage: "invalid learningObjectId request parameter",
+    paramsSchema: learningObjectIdSchema,
+  }),
+  getLearningObjectController,
+);
 
 /**
  * @route GET /learningObject/learningPath/:learningPathId
@@ -65,6 +77,10 @@ router.get("/:learningObjectId", protectAnyUser, getLearningObjectController);
 router.get(
   "/learningPath/:learningPathId",
   protectAnyUser,
+  validateRequest({
+    customErrorMessage: "invalid learningPathId request parameter",
+    paramsSchema: learningPathIdSchema,
+  }),
   getLearningObjectsForPathController,
 );
