--- dwengo_backend/zodSchemas/emailSchemas.ts
@@ -1,5 +1,0 @@
-import { z } from "zod";
-
-export const otherTeacherEmailBodySchema = z.object({
-  otherTeacherEmail: z.string().email(),
-});
