+++ dwengo_backend/zodSchemas/index.ts
@@ -1,0 +1,12 @@
+export * from "./body/authSchemas";
+export * from "./body/questionSchemas";
+export * from "./body/teamSchemas";
+export * from "./body/learningObjectSchemas";
+export * from "./body/commonSchemas";
+
+export * from "./params/commonSchemas";
+export * from "./params/numericIdSchemas";
+export * from "./params/stringIdSchemas";
+export * from "./params/mergedSchemas";
+
+export * from "./query/querySchemas";
