+++ dwengo_backend/zodSchemas/params/stringIdSchemas.ts
@@ -1,16 +1,5 @@
 import { z } from "zod";
-import { assignmentIdParamsSchema } from "./idSchemas";
 
-export const learningObjectParamsSchema = z.object({
-  hruid: z.string().min(1, "Hruid is required."),
-  language: z.string().min(1, "Language is required."),
-  version: z.coerce
-    .number()
-    .int()
-    .positive("Version must be a positive number."),
-});
-
-// Helper function to create string schemas
 const requiredId = (fieldName: string, label?: string) =>
   z.object({
     [fieldName]: z.string().min(1, `${label ?? fieldName} is required.`),
@@ -29,12 +18,9 @@ export const learningPathIdSchema = requiredId(
   "learningPathId",
   "LearningPathId",
 );
+export const learningObjectIdSchema = requiredId(
+  "learningObjectId",
+  "LearningObjectId",
+);
 export const nodeIdSchema = requiredId("nodeId", "NodeId");
 export const evaluationIdSchema = requiredId("evaluationId", "EvaluationId");
-
-export const nodeAndLearningPathIdSchema =
-  nodeIdSchema.merge(learningPathIdSchema);
-
-export const assignmentAndEvaluationIdParamSchema = evaluationIdSchema.merge(
-  assignmentIdParamsSchema,
-);
